WiP fix emergency mode
